src/backend_prediction/requirements.txt
joblib
fastapi
uvicorn
scikit-learn==1.1.2
kafka-python
